MOB-1171-Fixed creation of Transaction Item (#17)
* MOB-1171-Fixed creation of Transaction Item

* corrected parsing of id of transactionS

+++ unstoppable-ios-app/domains-manager-ios/NetworkEnvironment/MobileAPIAccess.swift
@@ -84,7 +84,7 @@ extension NetworkService {
     }
     
     struct TxResponse: Codable {
-        let id: Int
+        let id: UInt64
         let type: TxType
         let operation: TxOperation
         let statusGroup: String
@@ -774,7 +774,7 @@ extension NetworkService {
         }
         
         struct ActionsTxInfo: Decodable {
-            let id: UInt
+            let id: UInt64
             let type: String
             let blockchain: String
             let messageToSign: String?
@@ -787,7 +787,7 @@ extension NetworkService {
         }
         
         struct ActionsResponse: Decodable {
-            let id: UInt
+            let id: UInt64
             let domain: ActionsDomainInfo
             let txs: [ActionsTxInfo]
             let paymentInfo: ActionsPaymentInfo?
